+++ app/utils/scroll-into-view.js
@@ -1,3 +1,0 @@
-export {
-  default
-} from '@zestia/ember-select-box/utils/general/scroll-into-view';